♻️ refactor: [STRUCTURAL] 共有スキーマ型の互換 re-export を畳む
`chatAtom` / `selectionRects` / `ShelfPage` が `src/shared/schemas/` の型を
再輸出しており、CLAUDE.md が宣言する「front・server どちらにも同じ概念の型を
書かないこと」と実体がずれていた。輸入元をスキーマ本体へ寄せて alias を消す。

振る舞いは変えていない (jsdom 140 テストが変更前後で pass)。
