+++ src/front/components/ChatArea/ChatMessageBubble.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vite-plus/test";
 import { render, screen } from "@testing-library/react";
 import { ChatMessageBubble } from "./ChatMessageBubble";
-import type { ChatMessage } from "../../atoms/chatAtom";
+import type { ChatMessage } from "../../../shared/schemas/chat";
 
 function message(overrides: Partial<ChatMessage> = {}): ChatMessage {
   return {
