+++ src/front/components/ChatArea/ChatMessageList.tsx
@@ -1,6 +1,6 @@
 // oxlint-disable-next-line no-restricted-imports -- 新着メッセージとストリーム更新に追随して最下部へスクロールするために必要
 import { useEffect, useRef } from "react";
-import type { ChatMessage } from "../../atoms/chatAtom";
+import type { ChatMessage } from "../../../shared/schemas/chat";
 import { ChatMessageBubble } from "./ChatMessageBubble";
 
 interface ChatMessageListProps {
